# docs/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "TagResponse": {
        "description": "Response schema representing a normalized tag.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "TagResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Magti Internal Portal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/notifications/summary": {
      "get": {
        "description": "Combined bell-icon payload — unread mandatory readings + recent news + unread messages count.\n\nItem 15: the notifications popover used to call /my-readings + /news + /messages\nin sequence, each of which does its own DB roundtrip. This consolidates the\nfetch into ONE endpoint so the popover opens in roughly a single RTT.",
        "operationId": "get_notifications_summary_api_notifications_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Notifications Summary"
      }
    },
...
    "/api/tags": {
      "get": {
        "description": "Lists the normalized tag vocabulary (Block 1 tags/archiving foundation).\n\nTags are created on the fly by sync_tags() when an article/video is saved\nwith a non-empty tags field — there is no separate tag-creation endpoint.\n\nAccess: Authenticated users (any active role).",
        "operationId": "get_tags_api_tags_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TagResponse"
                  },
                  "title": "Response Get Tags Api Tags Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Tags"
      }
    },
...
  }
}